feat: Create HTML pages for three recipes
Create a new directory recipes with in the directory odin-recipes.
Create a new HTML file for each recipe: aglioEOlio.html,
kimchiFriedRice.html, and seaweedSoup.html
For each recipe HTML file, the boilerplate, the h1 heading, and the link to index.html are included.
At index.html, links for recipe HTML pages are included.