fix: edit img size for all recipe html
Adjusted the size of all images to be width=500px, height=auto.
It is because
1) uniformity of img sizes leads to consistency
and more readability of the pages.
2) when not setting any size values, if an image takes longer
to load than the rest of the page contents, it won’t take up
any space on the page at first but will
suddenly cause a drastic shift of the other page contents
once it does load in. Explicitly stating a height
and width prevents this from happening,
as space will be “reserved” on the page and appear
blank until the image loads.